Fix session_id assignation
Wrong implementation: the session_id is not the first sent cookie, but
the first key exchange hash.

As this value is not aimed to change, it is hard-set (self._session_id)
when the value becomes available, but one can't set it again (so use
self.session_id anywhere else to read it).