Improve authentication point
The `authenticate` function is the one that send all UserauthRequests.
If a password is provided, try with the password; however, if nothing is
provided, use a 'none' authentication.

One can use thie method even if we don't know supported authentication
methods by the server: if we use a wrong one, server will complain and
give us the correct list of supported methods.